requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
pyarrow==16.1.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
pydantic==2.7.4
